chore(auth): strip debug console output, extract decodeToken (#544)
User-approved cleanup (the file-length campaign's one sanctioned
behavior-adjacent change) to shrink the last `make file-length` offender,
`src/lib/stores/auth.svelte.ts` (580 lines).

- Remove 23 `[AUTH STORE]`-prefixed debug console.log flow-tracing statements.
  Keep the 9 genuine failure-reporting console.error/warn sites (HTTP refresh
  failure, malformed refresh response, ad-blocker/network diagnostic, permissions
  fetch failure, undecodable-token warn, and the two auto-refresh `.catch`
  handlers).
- Extract the pure `decodeToken` JWT helper to co-located `src/lib/stores/jwt.ts`
  (byte-identical body) and delegate to it.

Consequential lint-forced edits (0-warnings gate, no behavior change): the now
side-effect-free `catch (error)` in initialize() becomes `catch`, and two
orphaned display-only consts (`expiresInMinutes`, `refreshInMinutes`, consumed
only by removed logs) are dropped. Timers, race guards, and the in-memory-token
invariant are untouched.

Result: 580 → 517 lines. Still 17 over the 500 cap — the two sanctioned moves
cannot clear it (floor ~506 even stripping all logs); cap-vs-split is deferred
to the controller.
